pyproject.toml
[project]
name = "nys-design-optimisation-using-pysam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.4",
    "dagshub>=0.6.3",
    "deap>=1.4.3",
    "dvc>=3.64.2",
    # "matlabengine>=25.2.2",
    "mlflow>=3.6.0",
    "nlopt>=2.9.1",
    "nrel-pysam>=7.1.0",
    "openpyxl>=3.1.5",
    "optuna>=4.6.0",
    "pygad>=3.5.0",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
    "stable-baselines3>=2.7.1",
    "trackio>=0.15.0",
]
